Add interface for SchedulingAlgorithm.
